Moved unit tests for inventory-config API into a separate class
